fix : change route parameter for student pages
As the backend requires the student email to send data, the email was
being used as parameter for dynamic routing. The routes were something
like this: '/alunos/fulano.silva@edge.ufal.br/dados'.

However, this is not very practical for a URL. Instead, we have adopted
username as the parameter. The username is just the part to the left
of the @ in the student email. This results in much more readable URLS,
like: '/alunos/fulano.silva/dados'.

The downside is that the backend still requires the email for student
identification, so when a component receives the username parameter,
it must add the '@edge.ufal.br' at the end before using it to fetch
student data.